Move forwarder config to forwarder.

+++ src/SeqCli/Config/SeqCliStorageConfig.cs
@@ -2,5 +2,5 @@ namespace SeqCli.Config;
 
 public class SeqCliStorageConfig
 {
-    public long BufferSizeBytes { get; set; } = 67_108_864;
+    public int BufferSizeBytes { get; set; } = 67_108_864;
 }
